GH-809: simplify N4HeadlessCompiler by moving logic to BuildSetComputer
